Re-structure package names to NDK
