--- java/sample/src/main/java/io/sentry/ndk/samples/android/NativeSample.java
@@ -1,11 +1,0 @@
-package io.sentry.ndk.samples.android;
-
-public class NativeSample {
-  public static native void crash();
-
-  public static native void message();
-
-  static {
-    System.loadLibrary("native-sample");
-  }
-}
