--- java/settings.gradle
@@ -1,15 +1,0 @@
-pluginManagement {
-    repositories {
-        mavenCentral()
-        gradlePluginPortal()
-    }
-}
-
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-
-rootProject.name = "sentry-native-java"
-rootProject.buildFileName = "build.gradle.kts"
-
-include(
-    "lib", "sample"
-)
