+++ ndk/build.gradle.kts
