+++ ndk/debug.keystore
