+++ ndk/detekt.yml
