+++ ndk/gradle/wrapper/gradle-wrapper.properties
