+++ ndk/lib/CMakeLists.txt
