+++ ndk/lib/proguard-rules.pro
