+++ ndk/lib/src/main/java/io/sentry/ndk/DebugImage.java
