+++ ndk/sample/proguard-rules.pro
