+++ ndk/sample/src/main/AndroidManifest.xml
