+++ ndk/sample/src/main/java/io/sentry/ndk/sample/NdkSample.java
@@ -1,0 +1,11 @@
+package io.sentry.ndk.sample;
+
+public class NdkSample {
+    static {
+        System.loadLibrary("ndk-sample");
+    }
+
+    public static native void crash();
+
+    public static native void message();
+}
